Internal: Add optional escaping to k-v control [ED-20438]
