+++ packages/packages/libs/editor-controls/src/controls/key-value-control.tsx
@@ -13,6 +13,7 @@ import { __ } from '@wordpress/i18n';
 
 import { PropKeyProvider, PropProvider, useBoundProp } from '../bound-prop-context';
 import { createControl } from '../create-control';
+import { escapeHtmlAttr } from '../utils/escape-html-attr';
 import { TextControl } from './text-control';
 
 type KeyValueControlProps = {
@@ -21,6 +22,7 @@ type KeyValueControlProps = {
 	regexKey?: string;
 	regexValue?: string;
 	validationErrorMessage?: string;
+	escapeHtml?: boolean;
 	getHelperText?: ( key: string, value: string ) => { keyHelper?: string; valueHelper?: string };
 };
 
@@ -115,7 +117,11 @@ export const KeyValueControl = createControl( ( props: KeyValueControlProps = {}
 						{ keyLabel }
 					</FormLabel>
 					<PropKeyProvider bind={ 'key' }>
-						<TextControl inputValue={ sessionState.key } error={ !! keyError } helperText={ keyHelper } />
+						<TextControl
+							inputValue={ props.escapeHtml ? escapeHtmlAttr( sessionState.key ) : sessionState.key }
+							error={ !! keyError }
+							helperText={ keyHelper }
+						/>
 					</PropKeyProvider>
 					{ !! keyError && <FormHelperText error>{ keyError }</FormHelperText> }
 				</Grid>
@@ -125,7 +131,7 @@ export const KeyValueControl = createControl( ( props: KeyValueControlProps = {}
 					</FormLabel>
 					<PropKeyProvider bind={ 'value' }>
 						<TextControl
-							inputValue={ sessionState.value }
+							inputValue={ props.escapeHtml ? escapeHtmlAttr( sessionState.value ) : sessionState.value }
 							error={ !! valueError }
 							inputDisabled={ !! keyError }
 							helperText={ valueHelper }
