+++ packages/packages/libs/editor-controls/src/utils/escape-html-attr.ts
@@ -1,0 +1,11 @@
+export const escapeHtmlAttr = ( value: string ): string => {
+	const specialChars: Record< string, string > = {
+		'&': '&amp;',
+		'<': '&lt;',
+		'>': '&gt;',
+		"'": '&#39;',
+		'"': '&quot;',
+	};
+
+	return value.replace( /[&<>'"]/g, ( char ) => specialChars[ char ] || char );
+};
